This version replaces the Hull operation by a combinatorial operation that does not involve quotiens by binomials.  It does so by detecting embedded lattices and throwing in the corresponding monomials.  This idea is based on knowledge about mesoprimary decompositions and more features like this will be implemented in future versions.
First release in the 0.7 branch.

It is not entirely clear whether the new procedure is faster.  It should work best if there are many
embedded lattices and the old "minimalPrimaryComponent" function would call itself recursively many
times.  The benchmarks give confusing results anyway, e.g. performance on the third ideal has gotten
much worse lately, even if nothing about the Macaulay2 installation changed.

This new approach also circumenvents a problem where in the old version one would need to saturate a
non-pure difference lattice ideal.  The problem is demonstrated in the following example:

R = QQ[x,y,z,a,b]

I = ideal (x*y*(a^2-1),
           x*z*(b^2-1),
	   y*z*(a^2-b^2),
	   x*y*z*(a-1), x*y*z*(b-1),
	   x^2,y^2,z^2)
BPD I

In the old version there would arise a call to the saturate function which was failing because
I+P^(b) was not pure difference but had embedded non-saturated partial characters to be saturated.
The new approach avoids this problem since it directly removes embedded lattices without saturating
them.

For now the old minimalPrimaryComponent function was renamed to minimalPrimaryComponent2 and
survives in the source code.