Rename removeRedundant to extractInclusionMinimalIdeals
The former function removeRedundant claimed it would remove all
redundant ideals from an intersection, but this was not true.  It was
rewritten more efficiently and with a correct description of what it
does.  A test was installed and the documentation was updated

The unused function removeEmbedded was removed as it just duplicates
functionality.